+++ LibCarla/source/carla/trafficmanager/LocalizationStage.cpp
@@ -20,6 +20,9 @@ namespace LocalizationConstants {
   static const float MINIMUM_LANE_CHANGE_DISTANCE = 50.0f;
   static const float MAXIMUM_LANE_OBSTACLE_CURVATURE = 0.6f;
   static const uint64_t UNREGISTERED_ACTORS_SCAN_INTERVAL = 10;
+  static const float BLOCKED_TIME_THRESHOLD = 90.0f;
+  static const float DELTA_TIME_BETWEEN_DESTRUCTIONS = 10.0f;
+  static const float STOPPED_VELOCITY_THRESHOLD = 0.4f;  // meters per second.
 
 } // namespace LocalizationConstants
 
@@ -54,6 +57,12 @@ namespace LocalizationConstants {
 
     // Initializing the registered actors container state.
     registered_actors_state = -1;
+
+    // Initializing buffer lists.
+    buffer_list = std::make_shared<BufferList>();
+
+    // Initializing maximum idle time to null.
+    maximum_idle_time = std::make_pair(nullptr, 0.0);
   }
 
   LocalizationStage::~LocalizationStage() {}
@@ -68,6 +77,8 @@ namespace LocalizationConstants {
     const auto current_traffic_light_frame =
         traffic_light_frame_selector ? traffic_light_frame_a : traffic_light_frame_b;
 
+    // Selecting current timestamp from the world snapshot.
+    current_timestamp = world.GetSnapshot().GetTimestamp();
 
     // Looping over registered actors.
     for (uint64_t i = 0u; i < actor_list.size(); ++i) {
@@ -77,17 +88,19 @@ namespace LocalizationConstants {
       const cg::Location vehicle_location = vehicle->GetLocation();
       const float vehicle_velocity = vehicle->GetVelocity().Length();
 
-      //TODO: Improve search so it doesn't do it every loop..
-      auto search = idle_time.find(actor_id);
-      if (search == idle_time.end()) {
-        idle_time[actor_id] = chr::system_clock::now();
+      // Initializing idle times.
+      if (idle_time.find(actor_id) == idle_time.end() && current_timestamp.elapsed_seconds != 0) {
+        idle_time[actor_id] = current_timestamp.elapsed_seconds;
       }
 
       const float horizon_size = std::max(
           WAYPOINT_TIME_HORIZON * std::sqrt(vehicle_velocity * 10.0f),
           MINIMUM_HORIZON_LENGTH);
 
-      Buffer &waypoint_buffer = buffer_list->at(i);
+      if (buffer_list->find(actor_id) == buffer_list->end()) {
+        buffer_list->insert({actor_id, Buffer()});
+      }
+      Buffer &waypoint_buffer = buffer_list->at(actor_id);
 
       // Purge passed waypoints.
       if (!waypoint_buffer.empty()) {
@@ -198,9 +211,8 @@ namespace LocalizationConstants {
       }
 
       bool approaching_junction = false;
-      if (waypoint_buffer.front()->CheckJunction() ||
-          (look_ahead_point->CheckJunction() && !(waypoint_buffer.front()->CheckJunction()))) {
-        if (speed_limit > HIGHWAY_SPEED) {
+      if (look_ahead_point->CheckJunction() && !(waypoint_buffer.front()->CheckJunction())) {
+        if (speed_limit > HIGHWAY_SPEED*3.6f) {
           for (uint64_t j = 0u; (j < look_ahead_index) && !approaching_junction; ++j) {
             SimpleWaypointPtr swp = waypoint_buffer.at(j);
             if (swp->GetNextWaypoint().size() > 1) {
@@ -209,6 +221,24 @@ namespace LocalizationConstants {
           }
         } else {
           approaching_junction = true;
+        }
+      }
+
+      // Reset the variables when no longer approaching an intersection
+      if (!approaching_junction && approached[actor_id]){
+        final_safe_points[actor_id] = nullptr;
+        approached[actor_id] = false;
+      }
+
+      // Only do once, when the intersection has just been seen.
+      else if (approaching_junction && !approached[actor_id]){
+
+        SimpleWaypointPtr final_point = nullptr;
+        final_point = GetSafeLocationAfterJunction(vehicle_reference, waypoint_buffer);
+        if(final_point != nullptr){
+          final_safe_points[actor_id] = final_point;
+          approaching_junction = false;
+          approached[actor_id] = true;
         }
       }
 
@@ -232,6 +262,7 @@ namespace LocalizationConstants {
           actor_ptr = unregistered_actors.at(overlapping_actor_id);
         }
         collision_message.overlapping_actors.insert({overlapping_actor_id, actor_ptr});
+        collision_message.safe_point_after_junction = final_safe_points[actor_id];
       }
       collision_message.closest_waypoint = waypoint_buffer.front();
       collision_message.junction_look_ahead_waypoint = waypoint_buffer.at(look_ahead_index);
@@ -240,7 +271,17 @@ namespace LocalizationConstants {
       traffic_light_message.actor = vehicle;
       traffic_light_message.closest_waypoint = waypoint_buffer.front();
       traffic_light_message.junction_look_ahead_waypoint = waypoint_buffer.at(look_ahead_index);
+
+      // Updating idle time when necessary.
+      UpdateIdleTime(vehicle);
     }
+
+    if (IsVehicleStuck(maximum_idle_time.first)) {
+      TryDestroyVehicle(maximum_idle_time.first);
+    }
+
+    // Updating maximum idle time to null for the next iteration.
+    maximum_idle_time = std::make_pair(nullptr, current_timestamp.elapsed_seconds);
   }
 
   void LocalizationStage::DataReceiver() {
@@ -252,6 +293,7 @@ namespace LocalizationConstants {
 
       actor_list = registered_actors.GetList();
       uint64_t index = 0u;
+      vehicle_id_to_index.clear();
       for (auto &actor: actor_list) {
         vehicle_id_to_index.insert({actor->GetId(), index});
         ++index;
@@ -263,8 +305,7 @@ namespace LocalizationConstants {
     if (number_of_vehicles != actor_list.size()) {
 
       number_of_vehicles = static_cast<uint64_t>(actor_list.size());
-      // Allocating the buffer lists.
-      buffer_list = std::make_shared<BufferList>(number_of_vehicles);
+
       // Allocating output frames to be shared with the motion planner stage.
       planner_frame_a = std::make_shared<LocalizationToPlannerFrame>(number_of_vehicles);
       planner_frame_b = std::make_shared<LocalizationToPlannerFrame>(number_of_vehicles);
@@ -293,8 +334,12 @@ namespace LocalizationConstants {
 
   void LocalizationStage::DrawBuffer(Buffer &buffer) {
 
-    for (uint64_t i = 0u; i < buffer.size() && i < 5; ++i) {
-      debug_helper.DrawPoint(buffer.at(i)->GetLocation(), 0.1f, {255u, 0u, 0u}, 0.5f);
+    for (uint64_t i = 0u; i < buffer.size(); ++i) {
+      if(buffer.at(i)->GetWaypoint()->IsJunction()){
+        debug_helper.DrawPoint(buffer.at(i)->GetLocation() + cg::Location(0.0f,0.0f,2.0f), 0.3f, {0u, 0u, 255u}, 0.05f);
+      } else {
+        debug_helper.DrawPoint(buffer.at(i)->GetLocation() + cg::Location(0.0f,0.0f,2.0f), 0.3f, {0u, 255u, 255u}, 0.05f);
+      }
     }
   }
 
@@ -340,28 +385,30 @@ namespace LocalizationConstants {
     }
 
     // Regularly update unregistered actors.
-    std::vector<ActorId> actor_ids_to_erase;
-    for (auto& actor_info: unregistered_actors) {
-      if (actor_info.second->IsAlive()) {
-        cg::Location actor_location = actor_info.second->GetLocation();
+    const auto current_snapshot = world.GetSnapshot();
+    for (auto it = unregistered_actors.cbegin(); it != unregistered_actors.cend();) {
+      if (registered_actors.Contains(it->first) || !current_snapshot.Contains(it->first)) {
+        track_traffic.DeleteActor(it->first);
+        it = unregistered_actors.erase(it);
+      } else {
+        // Updating data structures.
+        cg::Location location = it->second->GetLocation();
+        const auto type = it->second->GetTypeId();
+
         SimpleWaypointPtr nearest_waypoint = nullptr;
-        const auto unregistered_type = actor_info.second->GetTypeId();
-        if (unregistered_type[0] == 'v') {
-          nearest_waypoint = local_map.GetWaypointInVicinity(actor_location);
-        } else if (unregistered_type[0] == 'w') {
-          nearest_waypoint = local_map.GetPedWaypoint(actor_location);
+        if (type[0] == 'v') {
+          nearest_waypoint = local_map.GetWaypointInVicinity(location);
+        } else if (type[0] == 'w') {
+          nearest_waypoint = local_map.GetPedWaypoint(location);
         }
         if (nearest_waypoint == nullptr) {
-          nearest_waypoint = local_map.GetWaypoint(actor_location);
+          nearest_waypoint = local_map.GetWaypoint(location);
         }
-        track_traffic.UpdateUnregisteredGridPosition(actor_info.first, nearest_waypoint);
-      } else {
-        track_traffic.DeleteActor(actor_info.first);
-        actor_ids_to_erase.push_back(actor_info.first);
+
+        track_traffic.UpdateUnregisteredGridPosition(it->first, nearest_waypoint);
+
+        ++it;
       }
-    }
-    for (auto actor_id: actor_ids_to_erase) {
-      unregistered_actors.erase(actor_id);
     }
   }
 
@@ -372,7 +419,7 @@ namespace LocalizationConstants {
     const float vehicle_velocity = vehicle->GetVelocity().Length();
     const float speed_limit = boost::static_pointer_cast<cc::Vehicle>(vehicle)->GetSpeedLimit();
 
-    const Buffer& waypoint_buffer = buffer_list->at(vehicle_id_to_index.at(actor_id));
+    const Buffer& waypoint_buffer = buffer_list->at(actor_id);
     const SimpleWaypointPtr& current_waypoint = waypoint_buffer.front();
 
     bool need_to_change_lane = false;
@@ -393,7 +440,10 @@ namespace LocalizationConstants {
         // This is totally ignoring unregistered actors during lane changes.
         // Need to fix this. Only a temporary solution.
         if (vehicle_id_to_index.find(other_vehicle_id) != vehicle_id_to_index.end()) {
-          const Buffer& other_buffer = buffer_list->at(vehicle_id_to_index.at(other_vehicle_id));
+          if (buffer_list->find(other_vehicle_id) == buffer_list->end()) {
+            buffer_list->insert({other_vehicle_id, Buffer()});
+          }
+          const Buffer& other_buffer = buffer_list->at(other_vehicle_id);
 
           if (!other_buffer.empty()) {
             const SimpleWaypointPtr& other_current_waypoint = other_buffer.front();
@@ -447,7 +497,7 @@ namespace LocalizationConstants {
       need_to_change_lane = true;
     }
 
-    const float change_over_distance = std::max(2.0f*vehicle_velocity, 10.0f);
+    const float change_over_distance =  cg::Math::Clamp(1.5f*vehicle_velocity, 3.0f, 20.0f);
     bool possible_to_lane_change = false;
     SimpleWaypointPtr change_over_point = nullptr;
 
@@ -487,10 +537,166 @@ namespace LocalizationConstants {
              !change_over_point->CheckJunction()) {
         change_over_point = change_over_point->GetNextWaypoint()[0];
       }
+
+      // Reset this variable if needed
+      if (approached[actor_id]){
+        approached[actor_id] = false;
+      }
+
       return change_over_point;
     } else {
       return nullptr;
     }
+  }
+
+SimpleWaypointPtr LocalizationStage::GetSafeLocationAfterJunction(const Vehicle &vehicle, Buffer &waypoint_buffer){
+
+    // Get the length of the car
+    float length = vehicle->GetBoundingBox().extent.x;
+
+    // First Waypoint before the junction
+    const SimpleWaypointPtr initial_point;
+    uint initial_index = 0;
+    // First Waypoint after the junction
+    SimpleWaypointPtr safe_point = nullptr;
+    uint safe_index = 0;
+    // Vehicle position after the junction
+    SimpleWaypointPtr final_point = nullptr;
+    // Safe space after the junction
+    const float safe_distance = 1.5f*length;
+
+    for (uint j = 0u; j < waypoint_buffer.size(); ++j){
+      if (waypoint_buffer.at(j)->CheckJunction()){
+        initial_index = j;
+        break;
+      }
+    }
+
+    // Stop if something failed
+    if (initial_index == 0 && !waypoint_buffer.front()->CheckJunction()){
+      return final_point;
+    }
+
+    // 2) Search for the end of the intersection (if it is in the buffer)
+    for (uint i = initial_index; i < waypoint_buffer.size(); ++i){
+
+      if (!waypoint_buffer.at(i)->CheckJunction()){
+        safe_point = waypoint_buffer.at(i);
+        safe_index = i;
+        break;
+      }
+    }
+
+    // If it hasn't been found, extend the buffer
+    if(safe_point == nullptr){
+      while (waypoint_buffer.back()->CheckJunction()) {
+
+          std::vector<SimpleWaypointPtr> next_waypoints = waypoint_buffer.back()->GetNextWaypoint();
+          uint selection_index = 0u;
+          if (next_waypoints.size() > 1) {
+            selection_index = static_cast<uint>(rand()) % next_waypoints.size();
+          }
+
+          waypoint_buffer.push_back(next_waypoints.at(selection_index));
+        }
+      // Save the last one
+      safe_point = waypoint_buffer.back();
+    }
+
+    // Stop if something failed
+    if (safe_index == 0){
+      return final_point;
+    }
+
+    // 3) Search for final_point (again, if it is in the buffer)
+  
+    for(uint k = safe_index; k < waypoint_buffer.size(); ++k){
+
+      if(safe_point->Distance(waypoint_buffer.at(k)->GetLocation()) > safe_distance){
+        final_point = waypoint_buffer.at(k);
+        break;
+      }
+    }
+
+    // If it hasn't been found, extend the buffer
+    if(final_point == nullptr){
+      while (safe_point->Distance(waypoint_buffer.back()->GetLocation()) < safe_distance) {
+
+        // Record the last point as a safe one and save it
+        std::vector<SimpleWaypointPtr> next_waypoints = waypoint_buffer.back()->GetNextWaypoint();
+        uint selection_index = 0u;
+        // Pseudo-randomized path selection if found more than one choice.
+        if (next_waypoints.size() > 1) {
+          selection_index = static_cast<uint>(rand()) % next_waypoints.size();
+        }
+
+        waypoint_buffer.push_back(next_waypoints.at(selection_index));
+      }
+      final_point = waypoint_buffer.back();
+    }
+
+    return final_point;
+  }
+
+  void LocalizationStage::UpdateIdleTime(const Actor& actor) {
+    if (idle_time.find(actor->GetId()) == idle_time.end()) {
+      return;
+    }
+
+    const auto vehicle = boost::static_pointer_cast<cc::Vehicle>(actor);
+    if (actor->GetVelocity().Length() > STOPPED_VELOCITY_THRESHOLD || (vehicle->IsAtTrafficLight() && vehicle->GetTrafficLightState() != TLS::Green)) {
+      idle_time[actor->GetId()] = current_timestamp.elapsed_seconds;
+    }
+
+    // Checking maximum idle time.
+    if (maximum_idle_time.first == nullptr || maximum_idle_time.second > idle_time[actor->GetId()]) {
+      maximum_idle_time = std::make_pair(actor, idle_time[actor->GetId()]);
+    }
+  }
+
+  bool LocalizationStage::IsVehicleStuck(const Actor& actor) {
+    if (actor == nullptr) {
+      return false;
+    }
+
+    if (idle_time.find(actor->GetId()) != idle_time.end()) {
+      auto delta_idle_time = current_timestamp.elapsed_seconds - idle_time.at(actor->GetId());
+      if (delta_idle_time >= BLOCKED_TIME_THRESHOLD) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  void LocalizationStage::CleanActor(const ActorId actor_id) {
+    track_traffic.DeleteActor(actor_id);
+    for (const auto& waypoint : buffer_list->at(actor_id)) {
+      track_traffic.RemovePassingVehicle(waypoint->GetId(), actor_id);
+    }
+
+    idle_time.erase(actor_id);
+    buffer_list->erase(actor_id);
+  }
+
+  bool LocalizationStage::TryDestroyVehicle(const Actor& actor) {
+    if (!actor->IsAlive()) {
+      return false;
+    }
+
+    const ActorId actor_id = actor->GetId();
+
+    auto delta_last_actor_destruction = current_timestamp.elapsed_seconds - elapsed_last_actor_destruction;
+    if (delta_last_actor_destruction >= DELTA_TIME_BETWEEN_DESTRUCTIONS) {
+      registered_actors.Destroy(actor);
+
+      // Clean actor data structures.
+      CleanActor(actor_id);
+
+      elapsed_last_actor_destruction = current_timestamp.elapsed_seconds;
+
+      return true;
+    }
+    return false;
   }
 
 } // namespace traffic_manager