+++ LibCarla/source/carla/trafficmanager/LocalizationStage.h
@@ -39,8 +39,10 @@ namespace traffic_manager {
   using namespace std::chrono;
   namespace cc = carla::client;
   using Actor = carla::SharedPtr<cc::Actor>;
+  using Vehicle = carla::SharedPtr<cc::Vehicle>;
   using ActorId = carla::ActorId;
   using ActorIdSet = std::unordered_set<ActorId>;
+  using TLS = carla::rpc::TrafficLightState;
 
   /// This class is responsible for maintaining a horizon of waypoints ahead
   /// of the vehicle for it to follow.
@@ -93,12 +95,18 @@ namespace traffic_manager {
     uint64_t number_of_vehicles;
     /// Used to only calculate the extended buffer once at junctions
     std::map<carla::ActorId, bool> approached;
-    /// Final Waypoint of the bounding box at intersections, amps to their respective IDs
-    std::map<carla::ActorId, SimpleWaypointPtr> final_points;
+    /// Point used to know if the junction has free space after its end, mapped to their respective actor id
+    std::map<carla::ActorId, SimpleWaypointPtr> final_safe_points;
     /// Object for tracking paths of the traffic vehicles.
     TrackTraffic track_traffic;
-    /// Map of all vehicles' idle time
-    std::unordered_map<ActorId, chr::time_point<chr::system_clock, chr::nanoseconds>> idle_time;
+    /// Map of all vehicles' idle time.
+    std::unordered_map<ActorId, double> idle_time;
+    /// Structure to hold the actor with the maximum idle time at each iteration.
+    std::pair<Actor, double> maximum_idle_time;
+    /// Variable to hold current timestamp from the world snapshot.
+    cc::Timestamp current_timestamp;
+    /// Simulated seconds since the beginning of the current episode when the last actor was destroyed.
+    double elapsed_last_actor_destruction = 0.0;
     /// Counter to track unregistered actors' scan interval.
     uint64_t unregistered_scan_duration = 0;
     /// A structure used to keep track of actors spawned outside of traffic
@@ -112,11 +120,22 @@ namespace traffic_manager {
 
     /// Method to determine lane change and obtain target lane waypoint.
     SimpleWaypointPtr AssignLaneChange(Actor vehicle, bool force, bool direction);
+
+    // When near an intersection, extends the buffer throughout all the
+    // intersection to see if there is space after it
+    SimpleWaypointPtr GetSafeLocationAfterJunction(const Vehicle &vehicle, Buffer &waypoint_buffer);
+
     /// Methods to modify waypoint buffer and track traffic.
     void PushWaypoint(Buffer& buffer, ActorId actor_id, SimpleWaypointPtr& waypoint);
     void PopWaypoint(Buffer& buffer, ActorId actor_id);
     /// Method to scan for unregistered actors and update their grid positioning.
     void ScanUnregisteredVehicles();
+
+    /// Methods for idle vehicle elimination.
+    void UpdateIdleTime(const Actor& actor);
+    bool IsVehicleStuck(const Actor& actor);
+    void CleanActor(const ActorId actor_id);
+    bool TryDestroyVehicle(const Actor& actor);
 
   public: